+++ packages/base/index.js
@@ -1,7 +1,7 @@
 // animations/
 import scroll from "./dist/animations/scroll.js";
 import slideDown from "./dist/animations/slideDown.js";
-import slideUp from "./dist/animations/slideup.js";
+import slideUp from "./dist/animations/slideUp.js";
 
 // config/
 import { getAnimationMode, setAnimationMode } from "./dist/config/AnimationMode.js";