+++ packages/main/src/Toast.ts
@@ -78,18 +78,6 @@ class Toast extends UI5Element {
 	/**
 	 * Defines the placement of the component.
 	 * <br><br>
-	 * Available options are:
-	 * <ul>
-	 * <li><code>TopStart</code></li>
-	 * <li><code>TopCenter</code></li>
-	 * <li><code>TopEnd</code></li>
-	 * <li><code>MiddleStart</code></li>
-	 * <li><code>MiddleCenter</code></li>
-	 * <li><code>MiddleEnd</code></li>
-	 * <li><code>BottomStart</code></li>
-	 * <li><code>BottomCenter</code></li>
-	 * <li><code>BottomEnd</code></li>
-	 * </ul>
 	 *
 	 * @type {sap.ui.webc.main.types.ToastPlacement}
 	 * @name sap.ui.webc.main.Toast.prototype.placement